hospital_comm_app/backend/requirements.txt
Flask
Flask_SQLAlchemy
Flask_Bcrypt
psycopg2
PyJWT

